moved linux serv manager back into ext-dep-setup
- added [Install] section to the .service file generation. (necessary for enabling services to work)
