+++ config/external-dep-setup.py
@@ -10,6 +10,7 @@ from io import BytesIO
 from pathlib import Path
 from zipfile import ZipFile
 from tarfile import TarFile
+from subprocess import PIPE
 
 import bs4
 import yaml
@@ -434,7 +435,151 @@ class LavalinkManager(Downloader):
         final_dest = lavalink_dir / "lavalink.jar"
         self.downloadURL(link, final_dest)
 
-#TODO finish implementing the WindowsDepednecyHandler and LinuxDependencyHandler abilities to download/install openjdk-17 and lavalink - then configure lavalink and create a service for it (in the case of windows)
+class LinuxServiceManager:
+    def is_in_project_root(self) -> bool:
+        """Checks if the terminal is in the ParkBot root directory."""
+        here = Path(os.getcwd())
+        root_components = {"cogs", "data"}
+        sub_dirs = set([x.name for x in here.iterdir() if x.is_dir()])
+        intersection = sub_dirs.intersection(root_components)
+        if root_components == intersection:
+            return True
+        return False
+    
+    def find_python(self) -> Path | None:
+        """Returns the path to the python executable in a virtual environment, on linux machines."""
+        parkbot_root = Path(os.getcwd())
+        desired_file = "python"
+        for dirpath, dirs, files in os.walk(parkbot_root):
+            for filename in files:
+                if (filename == desired_file):
+                    if os.access(Path(dirpath) / filename, os.X_OK):
+                        return Path(dirpath) / desired_file
+        return None    
+    
+    def __init__(self):
+        self.password = getpass.getpass("Please enter your linux user password:")
+        self.operating_sys = platform.system().lower()
+        if self.operating_sys != "linux":
+            raise OSError("This manager is only compatible with Linux systems.")
+        if not self.is_in_project_root():
+            raise FileNotFoundError("This script was not executed in the Parkbot root directory.")
+
+    def find_java(self) -> Path | None:
+        """Searches through a user's home directory and returns the first executable instance of 'java' which exists under a parent in 'acceptable_javas'."""
+        desired_file = "java"
+        user = getpass.getuser()
+        home_dir = Path(f"/home/{user}")
+        for dirpath, dirs, files in os.walk(home_dir):
+            for filename in files:
+                if (desired_file == filename):
+                    this_java = Path(dirpath) / desired_file
+                    if os.access(this_java, os.X_OK) and this_java.parent.parent.name in acceptable_javas:
+                        return Path(dirpath) / desired_file
+        return None
+
+    def find_lavalink_jar(self) -> Path | None:
+        """Returns the absolute path to the first instance of 'lavalink.jar' in a user's home directory. Recurses through all sub-dirs of home."""
+        desired_file = "lavalink.jar"
+        user = getpass.getuser()
+        dirs_to_search = [Path(f"/home/{user}")]
+        for directory in dirs_to_search:
+            for dirpath, dirs, files in os.walk(directory):
+                for filename in files:
+                    if (desired_file == filename):
+                        return Path(dirpath) / desired_file
+        return None
+
+    def generate_parkbot_service_file(self) -> None:
+        parkbot_root = Path(os.getcwd())
+        parkbot_python = self.find_python()
+        config = configparser.ConfigParser()
+        config.optionxform = str # preserve case
+        config["Unit"] = {
+            "Description": "A Discord bot service, bringing music and games to guilds it serves.",
+            "Before": "lavalink.service",
+            "Requires": "lavalink.service",
+        }
+        config["Service"] = {
+            "Type": "exec",
+            "WorkingDirectory": f"{parkbot_root}",
+            "ExecStart": f"{parkbot_python} {parkbot_root}/bot.py",
+            "Restart": "on-failure",
+            "RestartSec": "10",
+        }
+        config["Install"] ={
+            "WantedBy": "multi-user.target",
+        }
+
+        temp_dir = parkbot_root / "parkbot.service"
+        parkbot_service_path = Path("/etc/systemd/system/parkbot.service")
+        with open(temp_dir, "w") as file:
+            config.write(file)
+        process = subprocess.Popen(["sudo", "mv", str(temp_dir), str(parkbot_service_path)], stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        process.communicate(self.password.encode())
+        process.wait()
+
+    def generate_lavalink_service_file(self) -> None:
+        parkbot_root = Path(os.getcwd())
+        lavalink_jar = self.find_lavalink_jar() # both of these should 
+        java = self.find_java()
+        if not lavalink_jar or not java:
+            raise FileNotFoundError("Failed to create lavalink.service file. Could not find either `lavalink.jar` or the java 17 executable in this user's home directory.")
+        config = ConfigParser()
+        config.optionxform = str # preserve case
+        config["Unit"] = {
+            "Description": "Node for serving music to Discord",
+        }
+        config["Service"] = {
+            "Type": "exec",
+            "ExecStart": f"{java} -jar {lavalink_jar}",
+            "Restart": "on-failure",
+            "RestartSec": "10",
+        }
+        config["Install"] ={
+            "WantedBy": "multi-user.target",
+        }
+        
+        temp_dir = parkbot_root / "lavalink.service"
+        lavalink_service_path = Path("/etc/systemd/system/lavalink.service")
+        with open(temp_dir, "w") as file:
+            config.write(file)
+        process = subprocess.Popen(["sudo", "mv", str(temp_dir), str(lavalink_service_path)], stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        process.communicate(self.password.encode())
+        process.wait()
+
+    def enable_parkbot_service(self) -> None:
+        process = subprocess.Popen(["sudo", "systemctl", "enable", "parkbot.service"], stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        process.communicate(self.password.encode())
+        return_code = process.wait()
+        match return_code:
+            case 0: # success
+                print("Enabled parkbot.service.")
+            case 1:
+                print("Failed to enable parkbot.service.")
+            case _:
+                print("Failed to enable parkbot.service; unhandled return code.")
+
+    def enable_lavalink_service(self) -> None:
+        process = subprocess.Popen(["sudo", "systemctl", "enable", "lavalink.service"], stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        process.communicate(self.password.encode())
+        return_code = process.wait()
+        match return_code:
+            case 0: # success
+                print("Enabled lavalink.service.")
+            case 1:
+                print("Failed to enable lavalink.service.")
+            case _:
+                print("Failed to enable lavalink.service; unhandled return code.")
+
+    def generate_all_services(self) -> None:
+        self.generate_lavalink_service_file()
+        self.generate_parkbot_service_file()
+    
+    def enable_all_services(self) -> None:
+        self.enable_lavalink_service()
+        self.enable_parkbot_service()
+
 
 class ExternalDependencyHandler:
     """Class to model DependencyHandler behaviors and provide access to platform agnostic dependencies."""
@@ -628,6 +773,9 @@ class LinuxDependencyHandler(ExternalDependencyHandler):
         results['lavalink'] = lavalink_manager.findLavalink()
         results['lavalink_pass'] = lavalink_manager.configuration['lavalink']['server']['password']
         results['lavalink_uri'] = f"{lavalink_manager.configuration['server']['address']}:{lavalink_manager.configuration['server']['port']}"
+        service_manager = LinuxServiceManager()
+        service_manager.generate_all_services()
+        service_manager.enable_all_services()
         return results
     
     # def give_user_file_ownership(self, file:Path) -> bool:
